requirements.txt
flask==2.3.2
google-genai==0.16.0
gunicorn==20.1.0
